requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
brotli==1.2.0
brotlicffi==1.2.0.0
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
cryptography==46.0.3
cssselect2==0.8.0
dj-database-url==3.1.0
Django==5.2.10
django-allauth==65.13.1
django-bootstrap-modal-forms==3.0.5
django-bootstrap3==26.1
django-celery-beat==2.8.1
django-crispy-forms==2.5
django-csp==4.0
django-environ==0.12.0
django-forms-bootstrap==3.1.0
django-formtools==2.5.1
django-jazzmin==3.0.1
django-timezone-field==7.2.1
django-widget-tweaks==1.5.1
dotenv==0.9.9
fonttools==4.61.1
idna==3.11
joblib==1.5.3
kombu==5.6.2
nltk==3.9.2
packaging==25.0
pillow==12.1.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==2.23
pydyf==0.12.1
PyMySQL==1.1.2
pyphen==0.17.2
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
redis==7.1.0
regex==2025.11.3
reportlab==4.4.7
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.5
tinycss2==1.5.1
tinyhtml5==2.0.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2.14
weasyprint==67.0
webencodings==0.5.1
whitenoise==6.11.0
zopfli==0.4.0
django-jazzmin

